# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "Tasks_Manager app",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "This is a documentation of Personal Task's Manager app"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://task-manager-backend-y4a1.onrender.com"
    }
  ],
...
  "paths": {
...
    "/columns/:boardId": {
      "post": {
        "tags": [
          "Column"
        ],
        "summary": "Create column",
        "operationId": "createColumn",
        "description": "Create column with board Id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "67ca67e7ae7f10c88b598384"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "description": "column's name",
                    "type": "string",
                    "example": "To Do something"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Creates a board",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Column created successfully"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/columnSchema"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/columns/:columnId": {
      "patch": {
        "tags": [
          "Column"
        ],
        "summary": "Update column",
        "operationId": "updateColumn",
        "description": "Update column with column Id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "67ca67e7ae7f10c88b598384"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "description": "column's name",
                    "type": "string",
                    "example": "To Do something New"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Creates a board",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Column updated successfully"
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/columnSchema"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "delete": {
        "tags": [
          "Column"
        ],
        "summary": "Delete column",
        "operationId": "deleteColumn",
        "description": "Delete column",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "example": "67ca67e7ae7f10c88b598384"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted column"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "columnSchema": {
        "required": [
          "_id",
          "title",
          "boardId"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "Unique identifier for the column",
            "example": "67305580e57ecb50c727d2ea"
          },
          "title": {
            "type": "string",
            "description": "Title of the column",
            "example": "To Do something more #1"
          },
          "boardId": {
            "type": "string",
            "description": "Unique identifier of the board this column belongs to",
            "example": "672babd7db3166fdaba54a5c"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time when the column was created",
            "example": "2024-11-10T06:41:04.611Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Date and time when the column was last updated",
            "example": "2024-11-10T06:41:04.611Z"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "responses": {
...
      "404": {
        "description": "Not Found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message",
                "data"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "NotFoundError"
                },
                "data": {
                  "type": "object",
                  "example": {
                    "message": "Not found"
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}